Renames rendering subpackage
