+++ src/pyrite/_rendering/base_renderable.py
