+++ src/pyrite/_rendering/camera_renderer/__init__.py
