+++ src/pyrite/_rendering/camera_renderer/camera_renderer.py
