+++ src/pyrite/_rendering/ortho_projection.py
