+++ src/pyrite/_rendering/rect_bounds.py
