+++ src/pyrite/_rendering/render_texture.py
