+++ src/pyrite/_rendering/sprite_renderer/__init__.py
