+++ src/pyrite/_rendering/sprite_renderer/sprite_renderer.py
