+++ tests/test_camera_service.py
@@ -5,7 +5,7 @@ import unittest
 from pygame import Rect, Vector3
 
 from pyrite.services import CameraService
-from pyrite.rendering import OrthoProjection
+from pyrite._rendering import OrthoProjection
 from pyrite.types.projection import Projection
 from pyrite.types.camera import CameraBase
 from pyrite.transform import TransformComponent, Transform
